🚚 api: move route to src
* and move assessmentSupervisorAuthorization to src
