+++ api/src/certification/enrolment/application/certification-candidate-route.js
@@ -1,9 +1,11 @@
 import JoiDate from '@joi/date';
 import BaseJoi from 'joi';
 const Joi = BaseJoi.extend(JoiDate);
+import { certificationCandidatesController as libCertificationCandidatesController } from '../../../../lib/application/certification-candidates/certification-candidates-controller.js';
 import { securityPreHandlers } from '../../../shared/application/security-pre-handlers.js';
 import { identifiersType } from '../../../shared/domain/types/identifiers-type.js';
 import { authorization } from '../../shared/application/pre-handlers/authorization.js';
+import { assessmentSupervisorAuthorization } from '../../shared/application/pre-handlers/session-supervisor-authorization.js';
 import { certificationCandidateController } from './certification-candidate-controller.js';
 
 const register = async function (server) {
@@ -166,6 +168,32 @@ const register = async function (server) {
           'Cette route est restreinte aux utilisateurs authentifiés',
           'Elle supprime un candidat de certification à la session.',
         ],
+      },
+    },
+    {
+      method: 'POST',
+      path: '/api/certification-candidates/{id}/authorize-to-start',
+      config: {
+        validate: {
+          params: Joi.object({
+            id: identifiersType.certificationCandidateId,
+          }),
+          payload: Joi.object({
+            'authorized-to-start': Joi.boolean().required(),
+          }),
+        },
+        pre: [
+          {
+            method: assessmentSupervisorAuthorization.verifyByCertificationCandidateId,
+            assign: 'authorizationCheck',
+          },
+        ],
+        handler: libCertificationCandidatesController.authorizeToStart,
+        notes: [
+          '- **Cette route est restreinte aux utilisateurs authentifiés**\n' +
+            "- Indiquer la présence d'un candidat pour permettre ou bloquer son entrée en session",
+        ],
+        tags: ['api', 'certification-candidates'],
       },
     },
     {
