+++ api/src/certification/session-management/application/session-live-alert-route.js
@@ -1,8 +1,8 @@
 import Joi from 'joi';
 
-import { assessmentSupervisorAuthorization } from '../../../../lib/application/preHandlers/session-supervisor-authorization.js';
 import { responseObjectErrorDoc } from '../../../../lib/infrastructure/open-api-doc/livret-scolaire/response-object-error-doc.js';
 import { identifiersType } from '../../../shared/domain/types/identifiers-type.js';
+import { assessmentSupervisorAuthorization } from '../../shared/application/pre-handlers/session-supervisor-authorization.js';
 import { sessionLiveAlertController } from './session-live-alert-controller.js';
 
 const register = async function (server) {
