+++ api/src/certification/shared/application/pre-handlers/session-supervisor-authorization.js
@@ -1,5 +1,5 @@
-import * as supervisorAccessRepository from '../../../src/certification/session-management/infrastructure/repositories/supervisor-access-repository.js';
-import * as requestResponseUtils from '../../../src/shared/infrastructure/utils/request-response-utils.js';
+import * as requestResponseUtils from '../../../../shared/infrastructure/utils/request-response-utils.js';
+import * as supervisorAccessRepository from '../../../session-management/infrastructure/repositories/supervisor-access-repository.js';
 
 const verifyByCertificationCandidateId = async function (
   request,
