+++ api/tests/certification/enrolment/acceptance/application/certification-candidate-route_test.js
@@ -1,4 +1,4 @@
-import { CertificationCandidate } from '../../../../../lib/domain/models/CertificationCandidate.js';
+import { CertificationCandidate } from '../../../../../lib/domain/models/index.js';
 import { PIX_ADMIN } from '../../../../../src/authorization/domain/constants.js';
 import { ComplementaryCertificationKeys } from '../../../../../src/certification/shared/domain/models/ComplementaryCertificationKeys.js';
 import {
@@ -72,6 +72,60 @@ describe('Acceptance | Controller | Session | certification-candidate-route', fu
       expect(response.payload).to.equal(
         '{"data":[{"type":"certification-candidates","id":"1001","attributes":{"first-name":"first-name","last-name":"last-name","birthdate":"2000-01-04","birth-province-code":null,"birth-city":"PARIS 1","birth-country":"France","email":"somemail@example.net","result-recipient-email":"somerecipientmail@example.net","external-id":"externalId","extra-time-percentage":0.3,"is-linked":true,"organization-learner-id":null,"sex":"M","birth-insee-code":"75101","birth-postal-code":null,"complementary-certification":{"id":10000006,"label":"CléA Numérique","key":"CLEA"},"billing-mode":"PREPAID","prepayment-code":null}}]}',
       );
+    });
+  });
+
+  describe('POST /api/certification-candidates/:id/authorize-to-start', function () {
+    context('when user is authenticated', function () {
+      context('when the user is the supervisor of the session', function () {
+        it('should return a 204 status code', async function () {
+          // given
+          const server = await createServer();
+          const candidateUserId = databaseBuilder.factory.buildUser({}).id;
+          const certificationCenter = databaseBuilder.factory.buildCertificationCenter({});
+          const sessionId = databaseBuilder.factory.buildSession({
+            publishedAt: null,
+            certificationCenterId: certificationCenter.id,
+          }).id;
+          const certificationCourseId = databaseBuilder.factory.buildCertificationCourse({
+            sessionId,
+            userId: candidateUserId,
+          }).id;
+          const candidate = databaseBuilder.factory.buildCertificationCandidate({
+            id: 1001,
+            sessionId,
+            userId: candidateUserId,
+          });
+          databaseBuilder.factory.buildCoreSubscription({ certificationCandidateId: candidate.id });
+          databaseBuilder.factory.buildAssessment({
+            state: 'started',
+            userId: candidateUserId,
+            type: 'CERTIFICATION',
+            certificationCourseId,
+          });
+
+          const supervisorUserId = databaseBuilder.factory.buildUser({}).id;
+          databaseBuilder.factory.buildSupervisorAccess({
+            userId: supervisorUserId,
+            sessionId,
+          });
+
+          await databaseBuilder.commit();
+
+          const options = {
+            method: 'POST',
+            url: `/api/certification-candidates/${candidate.id}/authorize-to-start`,
+            headers: { authorization: generateValidRequestAuthorizationHeader(supervisorUserId, 'pix-certif') },
+            payload: { 'authorized-to-start': true },
+          };
+
+          // when
+          const response = await server.inject(options);
+
+          // then
+          expect(response.statusCode).to.equal(204);
+        });
+      });
     });
   });
