+++ api/tests/certification/enrolment/unit/application/certification-candidate-route_test.js
@@ -2,6 +2,7 @@ import { NotFoundError } from '../../../../../lib/domain/errors.js';
 import { certificationCandidateController } from '../../../../../src/certification/enrolment/application/certification-candidate-controller.js';
 import * as moduleUnderTest from '../../../../../src/certification/enrolment/application/certification-candidate-route.js';
 import { authorization } from '../../../../../src/certification/shared/application/pre-handlers/authorization.js';
+import { assessmentSupervisorAuthorization as sessionSupervisorAuthorization } from '../../../../../src/certification/shared/application/pre-handlers/session-supervisor-authorization.js';
 import { securityPreHandlers } from '../../../../../src/shared/application/security-pre-handlers.js';
 import { expect, HttpTestServer, sinon } from '../../../../test-helper.js';
 
@@ -271,6 +272,51 @@ describe('Unit | Application | Sessions | Routes', function () {
 
       // then
       expect(response.statusCode).to.equal(400);
+    });
+  });
+
+  describe('POST certification-candidates/{id}/authorize-to-start', function () {
+    it('should return 200 if the user is a supervisor of the session linked to the candidate', async function () {
+      //given
+      sinon
+        .stub(sessionSupervisorAuthorization, 'verifyByCertificationCandidateId')
+        .callsFake((request, h) => h.response(true));
+      sinon.stub(certificationCandidateController, 'authorizeToStart').returns('ok');
+
+      const httpTestServer = new HttpTestServer();
+      await httpTestServer.register(moduleUnderTest);
+      const payload = { 'authorized-to-start': true };
+
+      // when
+      const response = await httpTestServer.request(
+        'POST',
+        '/api/certification-candidates/1/authorize-to-start',
+        payload,
+      );
+
+      // then
+      expect(response.statusCode).to.equal(200);
+    });
+
+    it('should return 401 if the user is not a supervisor of the session linked to the candidate and certification center is in the whitelist', async function () {
+      //given
+      sinon
+        .stub(sessionSupervisorAuthorization, 'verifyByCertificationCandidateId')
+        .callsFake((request, h) => h.response().code(401).takeover());
+
+      const httpTestServer = new HttpTestServer();
+      await httpTestServer.register(moduleUnderTest);
+      const payload = { 'authorized-to-start': true };
+
+      // when
+      const response = await httpTestServer.request(
+        'POST',
+        '/api/certification-candidates/1/authorize-to-start',
+        payload,
+      );
+
+      // then
+      expect(response.statusCode).to.equal(401);
     });
   });
